+++ mod-data-import-converter-storage-server/src/main/java/org/folio/services/association/CommonProfileAssociationService.java
@@ -4,12 +4,12 @@ import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Promise;
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.folio.dao.ProfileDao;
 import org.folio.dao.association.MasterDetailAssociationDao;
 import org.folio.dao.association.ProfileAssociationDao;
-import org.folio.dataimport.util.OkapiConnectionParams;
+import org.folio.rest.impl.util.OkapiConnectionParams;
 import org.folio.rest.jaxrs.model.ActionProfile;
 import org.folio.rest.jaxrs.model.ActionProfileCollection;
 import org.folio.rest.jaxrs.model.JobProfile;
@@ -35,7 +35,7 @@ import java.util.UUID;
  */
 @Service
 public class CommonProfileAssociationService implements ProfileAssociationService {
-  private static final Logger LOGGER = LoggerFactory.getLogger(CommonProfileAssociationService.class);
+  private static final Logger LOGGER = LogManager.getLogger();
 
   @Autowired
   private ProfileDao<JobProfile, JobProfileCollection> jobProfileDao;