+++ NEWS.md
@@ -2,6 +2,7 @@
 * [MODDICONV-148](https://issues.folio.org/browse/MODDICONV-148) Create default OCLC Create Instance profile
 * [MODDICONV-149](https://issues.folio.org/browse/MODDICONV-149) Create default OCLC Update Instance profile
 * [MODDICONV-158](https://issues.folio.org/browse/MODDICONV-158) Update the names and descriptions of the default OCLC profiles.
+* [MODDICONV-160](https://issues.folio.org/browse/MODDICONV-160) Update the names and descriptions of the default OCLC match, action, field mapping profiles.
 
 ## 2020-11-06 v1.9.2
 * [MODDICONV-152](https://issues.folio.org/browse/MODDICONV-150) Upgrade to RMB v31.1.5